requirements.txt
asgiref==3.11.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
gunicorn==25.1.0
openpyxl==3.1.2
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.0.1
setuptools==82.0.0
sqlparse==0.5.5
tzdata==2025.3
wheel==0.46.3
whitenoise==6.12.0
